Remove event-structure.js and integrate conflict resolution into SharedCore
